pyproject.toml
[project]
name = "blockchain-llm-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "tjjd4"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "transformers (>=4.49.0,<5.0.0)",
    "gradio (>=5.20.1,<6.0.0)",
    "tf-keras (>=2.19.0,<3.0.0)",
    "huggingface-hub[cli] (>=0.29.3,<0.30.0)",
    "torch (>=2.6.0,<3.0.0)",
    "accelerate (>=1.5.1,<2.0.0)",
    "ollama (>=0.4.7,<0.5.0)",
    "openai (>=1.66.3,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]

[tool.poetry]
package-mode = false
packages = [{ include = ".", from = "." }]

[tool.poetry.group.dev.dependencies]
poetry-dotenv-plugin = "^0.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
